+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for Haskell
 
-## 0.1.0.0 -- YYYY-mm-dd
+## MVP1 -- 2025-07-11
 
-* First version. Released on an unsuspecting world.
+* Generation of a map with fixed size;
+* Random spawn of loot and enemies in the room.
+
+## MVP2 -- 2025-07-17
+* Opportunity to modify the parameters of the map
+* Opportunity to modify the chance of enemies and loot spawn
+* Improved generation of the map (conncectedness)